Add action caching in pages controller
Page caching can't be used on controller actions that have before filters so we can use action caching instead. Action caching allows the before filters to run before the cache is served. In the console you'll see that it is writing to the cache when you request the  page Write fragment (views/localhost:3000/restricted) then reading from it in the next request (Read fragment views/localhost:3000/restricted)